IoUring: Use correct native jni types in function definition (#14914)
Motivation:

When define functions that are exposed via JNI we need to ensure we use
the types that are provided and supported by JNI. Not doing so might
result in undefined behaviour depending on the platform.

Modification:

Fix netty_io_uring_unregister_buf_ring to use the correct supported
types.

Result:

No undefined behaviour anymore when unregister buffer ring

+++ transport-native-io_uring/src/test/java/io/netty/channel/uring/IoUringBufferRingTest.java
@@ -60,7 +60,7 @@ public class IoUringBufferRingTest {
             assumeTrue(
                     ioUringBufRingAddr > 0,
                     "ioUringSetupBufRing result must great than 0, but now result is " + ioUringBufRingAddr);
-            int freeRes = Native.ioUringUnRegisterBufRing(ringFd, ioUringBufRingAddr, 4, 1);
+            int freeRes = Native.ioUringUnRegisterBufRing(ringFd, ioUringBufRingAddr, 4, (short) 1);
             assertEquals(
                     0,
                     freeRes,
